refactor(frontend): extract shared PairingQr component
