+++ src/modals/pair-device.tsx
@@ -7,11 +7,11 @@
    in which case we show instructions rather than a useless QR. */
 
 import { useEffect, useState } from 'react';
-import QRCode from 'qrcode';
 
 import { api } from '../lib/api';
 import { IconClose, IconCopy, IconQrCode, IconShield, IconCheck } from '../lib/icons';
 import type { PairSessionInfo } from '../lib/types';
+import { PairingQr } from '../components/pairing/pairing-qr';
 
 interface PairDeviceModalProps {
   open: boolean;
@@ -21,14 +21,7 @@ interface PairDeviceModalProps {
 export function PairDeviceModal({ open, onClose }: PairDeviceModalProps) {
   const [info, setInfo] = useState<PairSessionInfo | null>(null);
   const [status, setStatus] = useState<'loading' | 'ready' | 'unavailable' | 'error'>('loading');
-  const [qrDataUrl, setQrDataUrl] = useState<string | null>(null);
   const [nonce, setNonce] = useState(0); // bump to regenerate the code
-  const [now, setNow] = useState(() => Date.now());
-  useEffect(() => {
-    if (status !== 'ready') return;
-    const id = window.setInterval(() => setNow(Date.now()), 1000);
-    return () => window.clearInterval(id);
-  }, [status]);
 
   /* Fetch a new pairing session each time the modal opens (or the user hits
      "Regenerate code"). Sessions are short-lived, so don't cache across opens. */
@@ -36,7 +29,6 @@ export function PairDeviceModal({ open, onClose }: PairDeviceModalProps) {
     if (!open) return;
     let cancelled = false;
     setStatus('loading');
-    setQrDataUrl(null);
     api
       .createPairSession()
       .then((r) => {
@@ -52,25 +44,6 @@ export function PairDeviceModal({ open, onClose }: PairDeviceModalProps) {
       cancelled = true;
     };
   }, [open, nonce]);
-
-  /* Render the QR from the compact payload once we have session info. */
-  useEffect(() => {
-    if (status !== 'ready' || !info) {
-      setQrDataUrl(null);
-      return;
-    }
-    let cancelled = false;
-    QRCode.toDataURL(info.qrPayload, { margin: 4, scale: 8, errorCorrectionLevel: 'M' })
-      .then((d) => {
-        if (!cancelled) setQrDataUrl(d);
-      })
-      .catch(() => {
-        if (!cancelled) setQrDataUrl(null);
-      });
-    return () => {
-      cancelled = true;
-    };
-  }, [status, info]);
 
   if (!open) return null;
 
@@ -147,43 +120,11 @@ export function PairDeviceModal({ open, onClose }: PairDeviceModalProps) {
                   In the app, tap <strong>Pair a device → Scan QR</strong> and point the camera here.
                   Your phone must be on the same Wi‑Fi.
                 </p>
-                <div className="grid place-items-center">
-                  <div className="bg-white p-3 rounded-2xl border border-ink/10">
-                    {qrDataUrl ? (
-                      <img
-                        src={qrDataUrl}
-                        alt="Pairing QR code"
-                        data-testid="pair-qr-image"
-                        width={288}
-                        height={288}
-                        className="block w-72 h-72"
-                        style={{ imageRendering: 'pixelated' }}
-                      />
-                    ) : (
-                      <div className="w-72 h-72 grid place-items-center text-ink/40">
-                        Generating…
-                      </div>
-                    )}
-                  </div>
-                </div>
-                <button
-                  onClick={() => setNonce((n) => n + 1)}
-                  className="text-xs text-magenta hover:underline min-h-[44px]"
-                >
-                  Regenerate code
-                </button>
-                {(() => {
-                  const remainingMs = Math.max(0, info.expiresAt - now);
-                  const mm = Math.floor(remainingMs / 60000);
-                  const ss = Math.floor((remainingMs % 60000) / 1000);
-                  return (
-                    <p data-testid="pair-code-countdown" className="text-xs text-ink/50">
-                      {remainingMs > 0
-                        ? `This code expires in ${mm}:${ss.toString().padStart(2, '0')}.`
-                        : 'This code has expired — tap Regenerate code.'}
-                    </p>
-                  );
-                })()}
+                <PairingQr
+                  payload={info.qrPayload}
+                  expiresAt={info.expiresAt}
+                  onRegenerate={() => setNonce((n) => n + 1)}
+                />
                 <details className="rounded-xl border border-ink/10 bg-ink/[0.02]">
                   <summary className="px-4 py-3 cursor-pointer text-ink/70 font-medium select-none">
                     Or enter these manually
